Align device types with Cocoa and Java SDKs (#3567)
