+++ src/Sentry.Maui/Internal/MauiDeviceData.cs
@@ -40,7 +40,7 @@ internal static class MauiDeviceData
             try
             {
                 var battery = Battery.Default;
-                device.BatteryLevel ??= battery.ChargeLevel < 0 ? null : (short)(battery.ChargeLevel * 100.0);
+                device.BatteryLevel ??= battery.ChargeLevel < 0 ? null : (float)(battery.ChargeLevel * 100.0);
                 device.BatteryStatus ??= battery.State.ToString();
                 device.IsCharging ??= battery.State switch
                 {
