+++ test/Sentry.Tests/ApiApprovalTests.Run.DotNet6_0.verified.txt
@@ -1588,7 +1588,7 @@ namespace Sentry.Protocol
         public const string Type = "device";
         public Device() { }
         public string? Architecture { get; set; }
-        public short? BatteryLevel { get; set; }
+        public float? BatteryLevel { get; set; }
         public string? BatteryStatus { get; set; }
         public System.DateTimeOffset? BootTime { get; set; }
         public string? Brand { get; set; }
@@ -1610,7 +1610,7 @@ namespace Sentry.Protocol
         public string? Name { get; set; }
         public Sentry.Protocol.DeviceOrientation? Orientation { get; set; }
         public int? ProcessorCount { get; set; }
-        public int? ProcessorFrequency { get; set; }
+        public float? ProcessorFrequency { get; set; }
         public float? ScreenDensity { get; set; }
         public int? ScreenDpi { get; set; }
         public string? ScreenResolution { get; set; }
