+++ test/Sentry.Tests/Protocol/Context/DeviceTests.cs
@@ -3,6 +3,7 @@ namespace Sentry.Tests.Protocol.Context;
 public class DeviceTests
 {
     private readonly IDiagnosticLogger _testOutputLogger;
+    private const float Delta = 0.0001f;
 
     public DeviceTests(ITestOutputHelper output)
     {
@@ -116,7 +117,129 @@ public class DeviceTests
     [Fact]
     public void Clone_CopyValues()
     {
-        var sut = new Device
+        var sut = TestDevice();
+
+        var clone = sut.Clone();
+
+        AssertAreEqual(sut, clone);
+    }
+
+    [Fact]
+    public void WriteTo_FromJson_Symmetric()
+    {
+        // Arrange
+        var sut = TestDevice();
+
+        var json = sut.ToJsonString();
+        using var document = JsonDocument.Parse(json);
+        var jsonElement = document.RootElement;
+
+        // Act
+        var result = Device.FromJson(jsonElement);
+
+        // Assert
+        AssertAreEqual(result, sut);
+    }
+
+    [Fact]
+    public void FromJson_NonSystemTimeZone_NoException()
+    {
+        // Arrange
+        const string json = """{"type":"device","timezone":"tz_id","timezone_display_name":"tz_name"}""";
+
+        // Act
+        var device = Json.Parse(json, Device.FromJson);
+
+        // Assert
+        device.Timezone.Should().NotBeNull();
+        device.Timezone?.Id.Should().Be("tz_id");
+        device.Timezone?.DisplayName.Should().Be("tz_name");
+    }
+
+    [Fact]
+    public void FromJson_BatteryLevelFloat()
+    {
+        // Arrange
+        const string json = """{"type":"device","battery_level":1.5}""";
+
+        // Act
+        var device = Json.Parse(json, Device.FromJson);
+
+        // Assert
+        device.BatteryLevel.Should().NotBeNull();
+        device.BatteryLevel?.Should().BeApproximately(1.5f, Delta);
+    }
+
+    [Fact]
+    public void FromJson_ProcessorFrequencyFloat()
+    {
+        // Arrange
+        const string json = """{"type":"device","processor_frequency":2500.3}""";
+
+        // Act
+        var device = Json.Parse(json, Device.FromJson);
+
+        // Assert
+        device.ProcessorFrequency.Should().NotBeNull();
+        device.ProcessorFrequency?.Should().BeApproximately(2500.3f, Delta);
+    }
+
+    [Theory]
+    [MemberData(nameof(TestCases))]
+    public void SerializeObject_TestCase_SerializesAsExpected((Device device, string serialized) @case)
+    {
+        var actual = @case.device.ToJsonString(_testOutputLogger);
+
+        Assert.Equal(@case.serialized, actual);
+    }
+
+    public static IEnumerable<object[]> TestCases()
+    {
+        yield return new object[] { (new Device(), """{"type":"device"}""") };
+        yield return new object[] { (new Device { Name = "some name" }, """{"type":"device","name":"some name"}""") };
+        yield return new object[] { (new Device { Orientation = DeviceOrientation.Landscape }, """{"type":"device","orientation":"landscape"}""") };
+        yield return new object[] { (new Device { Brand = "some brand" }, """{"type":"device","brand":"some brand"}""") };
+        yield return new object[] { (new Device { Manufacturer = "some manufacturer" }, """{"type":"device","manufacturer":"some manufacturer"}""") };
+        yield return new object[] { (new Device { Family = "some family" }, """{"type":"device","family":"some family"}""") };
+        yield return new object[] { (new Device { Model = "some model" }, """{"type":"device","model":"some model"}""") };
+        yield return new object[] { (new Device { ModelId = "some model id" }, """{"type":"device","model_id":"some model id"}""") };
+        yield return new object[] { (new Device { Architecture = "some arch" }, """{"type":"device","arch":"some arch"}""") };
+        yield return new object[] { (new Device { BatteryLevel = 1 }, """{"type":"device","battery_level":1}""") };
+        yield return new object[] { (new Device { BatteryLevel = 1.5f }, """{"type":"device","battery_level":1.5}""") };
+        yield return new object[] { (new Device { IsCharging = true }, """{"type":"device","charging":true}""") };
+        yield return new object[] { (new Device { IsOnline = true }, """{"type":"device","online":true}""") };
+        yield return new object[] { (new Device { Simulator = false }, """{"type":"device","simulator":false}""") };
+        yield return new object[] { (new Device { MemorySize = 1 }, """{"type":"device","memory_size":1}""") };
+        yield return new object[] { (new Device { FreeMemory = 1 }, """{"type":"device","free_memory":1}""") };
+        yield return new object[] { (new Device { UsableMemory = 1 }, """{"type":"device","usable_memory":1}""") };
+        yield return new object[] { (new Device { LowMemory = true }, """{"type":"device","low_memory":true}""") };
+        yield return new object[] { (new Device { StorageSize = 1 }, """{"type":"device","storage_size":1}""") };
+        yield return new object[] { (new Device { FreeStorage = 1 }, """{"type":"device","free_storage":1}""") };
+        yield return new object[] { (new Device { ExternalStorageSize = 1 }, """{"type":"device","external_storage_size":1}""") };
+        yield return new object[] { (new Device { ExternalFreeStorage = 1 }, """{"type":"device","external_free_storage":1}""") };
+        yield return new object[] { (new Device { ScreenResolution = "1x1" }, """{"type":"device","screen_resolution":"1x1"}""") };
+        yield return new object[] { (new Device { ScreenDensity = 1 }, """{"type":"device","screen_density":1}""") };
+        yield return new object[] { (new Device { ScreenDpi = 1 }, """{"type":"device","screen_dpi":1}""") };
+        yield return new object[] { (new Device { BootTime = DateTimeOffset.MaxValue }, """{"type":"device","boot_time":"9999-12-31T23:59:59.9999999+00:00"}""") };
+        yield return new object[] { (new Device { Timezone = TimeZoneInfo.CreateCustomTimeZone("tz_id", TimeSpan.Zero, "tz_name", "tz_name") }, """{"type":"device","timezone":"tz_id","timezone_display_name":"tz_name"}""") };
+        yield return new object[] { (new Device { Timezone = TimeZoneInfo.CreateCustomTimeZone("tz_id", TimeSpan.Zero, "tz_id", "tz_id") }, """{"type":"device","timezone":"tz_id"}""") };
+        yield return new object[] { (new Device { ProcessorCount = 8 }, """{"type":"device","processor_count":8}""") };
+        yield return new object[] { (new Device { CpuDescription = "Intel(R) Core(TM)2 Quad CPU Q6600 @ 2.40GHz" }, """{"type":"device","cpu_description":"Intel(R) Core(TM)2 Quad CPU Q6600 @ 2.40GHz"}""") };
+        yield return new object[] { (new Device { ProcessorFrequency = 2500 }, """{"type":"device","processor_frequency":2500}""") };
+        yield return new object[] { (new Device { ProcessorFrequency = 2500.5f }, """{"type":"device","processor_frequency":2500.5}""") };
+        yield return new object[] { (new Device { DeviceType = "Handheld" }, """{"type":"device","device_type":"Handheld"}""") };
+        yield return new object[] { (new Device { BatteryStatus = "Charging" }, """{"type":"device","battery_status":"Charging"}""") };
+        yield return new object[] { (new Device { DeviceUniqueIdentifier = "d610540d-11d6-4daa-a98c-b71030acae4d" }, """{"type":"device","device_unique_identifier":"d610540d-11d6-4daa-a98c-b71030acae4d"}""") };
+        yield return new object[] { (new Device { SupportsVibration = false }, """{"type":"device","supports_vibration":false}""") };
+        yield return new object[] { (new Device { SupportsAccelerometer = true }, """{"type":"device","supports_accelerometer":true}""") };
+        yield return new object[] { (new Device { SupportsGyroscope = true }, """{"type":"device","supports_gyroscope":true}""") };
+        yield return new object[] { (new Device { SupportsAudio = true }, """{"type":"device","supports_audio":true}""") };
+        yield return new object[] { (new Device { SupportsLocationService = true }, """{"type":"device","supports_location_service":true}""") };
+    }
+
+    private static Device TestDevice()
+    {
+        return new Device
         {
             Name = "name",
             Brand = "brand",
@@ -155,110 +278,48 @@ public class DeviceTests
             SupportsAudio = true,
             SupportsLocationService = true
         };
-
-        var clone = sut.Clone();
-
-        Assert.Equal(sut.Name, clone.Name);
-        Assert.Equal(sut.Family, clone.Family);
-        Assert.Equal(sut.Brand, clone.Brand);
-        Assert.Equal(sut.Manufacturer, clone.Manufacturer);
-        Assert.Equal(sut.Model, clone.Model);
-        Assert.Equal(sut.ModelId, clone.ModelId);
-        Assert.Equal(sut.Architecture, clone.Architecture);
-        Assert.Equal(sut.BatteryLevel, clone.BatteryLevel);
-        Assert.Equal(sut.IsCharging, clone.IsCharging);
-        Assert.Equal(sut.Orientation, clone.Orientation);
-        Assert.Equal(sut.Simulator, clone.Simulator);
-        Assert.Equal(sut.MemorySize, clone.MemorySize);
-        Assert.Equal(sut.FreeMemory, clone.FreeMemory);
-        Assert.Equal(sut.LowMemory, clone.LowMemory);
-        Assert.Equal(sut.UsableMemory, clone.UsableMemory);
-        Assert.Equal(sut.StorageSize, clone.StorageSize);
-        Assert.Equal(sut.FreeStorage, clone.FreeStorage);
-        Assert.Equal(sut.ExternalStorageSize, clone.ExternalStorageSize);
-        Assert.Equal(sut.ExternalFreeStorage, clone.ExternalFreeStorage);
-        Assert.Equal(sut.ScreenResolution, clone.ScreenResolution);
-        Assert.Equal(sut.ScreenDensity, clone.ScreenDensity);
-        Assert.Equal(sut.ScreenDpi, clone.ScreenDpi);
-        Assert.Equal(sut.BootTime, clone.BootTime);
-        Assert.Equal(sut.Timezone, clone.Timezone);
-        Assert.Equal(sut.IsOnline, clone.IsOnline);
-        Assert.Equal(sut.ProcessorCount, clone.ProcessorCount);
-        Assert.Equal(sut.CpuDescription, clone.CpuDescription);
-        Assert.Equal(sut.ProcessorFrequency, clone.ProcessorFrequency);
-        Assert.Equal(sut.DeviceType, clone.DeviceType);
-        Assert.Equal(sut.BatteryStatus, clone.BatteryStatus);
-        Assert.Equal(sut.DeviceUniqueIdentifier, clone.DeviceUniqueIdentifier);
-        Assert.Equal(sut.SupportsVibration, clone.SupportsVibration);
-        Assert.Equal(sut.SupportsAccelerometer, clone.SupportsAccelerometer);
-        Assert.Equal(sut.SupportsGyroscope, clone.SupportsGyroscope);
-        Assert.Equal(sut.SupportsAudio, clone.SupportsAudio);
-        Assert.Equal(sut.SupportsLocationService, clone.SupportsLocationService);
     }
 
-    [Theory]
-    [MemberData(nameof(TestCases))]
-    public void SerializeObject_TestCase_SerializesAsExpected((Device device, string serialized) @case)
+    private static void AssertAreEqual(Device actual, Device expected)
     {
-        var actual = @case.device.ToJsonString(_testOutputLogger);
-
-        Assert.Equal(@case.serialized, actual);
-    }
-
-    [Fact]
-    public void FromJson_NonSystemTimeZone_NoException()
-    {
-        // Arrange
-        const string json = """{"type":"device","timezone":"tz_id","timezone_display_name":"tz_name"}""";
-
-        // Act
-        var device = Json.Parse(json, Device.FromJson);
-
-        // Assert
-        device.Timezone.Should().NotBeNull();
-        device.Timezone?.Id.Should().Be("tz_id");
-        device.Timezone?.DisplayName.Should().Be("tz_name");
-    }
-
-    public static IEnumerable<object[]> TestCases()
-    {
-        yield return new object[] { (new Device(), """{"type":"device"}""") };
-        yield return new object[] { (new Device { Name = "some name" }, """{"type":"device","name":"some name"}""") };
-        yield return new object[] { (new Device { Orientation = DeviceOrientation.Landscape }, """{"type":"device","orientation":"landscape"}""") };
-        yield return new object[] { (new Device { Brand = "some brand" }, """{"type":"device","brand":"some brand"}""") };
-        yield return new object[] { (new Device { Manufacturer = "some manufacturer" }, """{"type":"device","manufacturer":"some manufacturer"}""") };
-        yield return new object[] { (new Device { Family = "some family" }, """{"type":"device","family":"some family"}""") };
-        yield return new object[] { (new Device { Model = "some model" }, """{"type":"device","model":"some model"}""") };
-        yield return new object[] { (new Device { ModelId = "some model id" }, """{"type":"device","model_id":"some model id"}""") };
-        yield return new object[] { (new Device { Architecture = "some arch" }, """{"type":"device","arch":"some arch"}""") };
-        yield return new object[] { (new Device { BatteryLevel = 1 }, """{"type":"device","battery_level":1}""") };
-        yield return new object[] { (new Device { IsCharging = true }, """{"type":"device","charging":true}""") };
-        yield return new object[] { (new Device { IsOnline = true }, """{"type":"device","online":true}""") };
-        yield return new object[] { (new Device { Simulator = false }, """{"type":"device","simulator":false}""") };
-        yield return new object[] { (new Device { MemorySize = 1 }, """{"type":"device","memory_size":1}""") };
-        yield return new object[] { (new Device { FreeMemory = 1 }, """{"type":"device","free_memory":1}""") };
-        yield return new object[] { (new Device { UsableMemory = 1 }, """{"type":"device","usable_memory":1}""") };
-        yield return new object[] { (new Device { LowMemory = true }, """{"type":"device","low_memory":true}""") };
-        yield return new object[] { (new Device { StorageSize = 1 }, """{"type":"device","storage_size":1}""") };
-        yield return new object[] { (new Device { FreeStorage = 1 }, """{"type":"device","free_storage":1}""") };
-        yield return new object[] { (new Device { ExternalStorageSize = 1 }, """{"type":"device","external_storage_size":1}""") };
-        yield return new object[] { (new Device { ExternalFreeStorage = 1 }, """{"type":"device","external_free_storage":1}""") };
-        yield return new object[] { (new Device { ScreenResolution = "1x1" }, """{"type":"device","screen_resolution":"1x1"}""") };
-        yield return new object[] { (new Device { ScreenDensity = 1 }, """{"type":"device","screen_density":1}""") };
-        yield return new object[] { (new Device { ScreenDpi = 1 }, """{"type":"device","screen_dpi":1}""") };
-        yield return new object[] { (new Device { BootTime = DateTimeOffset.MaxValue }, """{"type":"device","boot_time":"9999-12-31T23:59:59.9999999+00:00"}""") };
-        yield return new object[] { (new Device { Timezone = TimeZoneInfo.CreateCustomTimeZone("tz_id", TimeSpan.Zero, "tz_name", "tz_name") }, """{"type":"device","timezone":"tz_id","timezone_display_name":"tz_name"}""") };
-        yield return new object[] { (new Device { Timezone = TimeZoneInfo.CreateCustomTimeZone("tz_id", TimeSpan.Zero, "tz_id", "tz_id") }, """{"type":"device","timezone":"tz_id"}""") };
-        yield return new object[] { (new Device { ProcessorCount = 8 }, """{"type":"device","processor_count":8}""") };
-        yield return new object[] { (new Device { CpuDescription = "Intel(R) Core(TM)2 Quad CPU Q6600 @ 2.40GHz" }, """{"type":"device","cpu_description":"Intel(R) Core(TM)2 Quad CPU Q6600 @ 2.40GHz"}""") };
-        yield return new object[] { (new Device { ProcessorFrequency = 2500 }, """{"type":"device","processor_frequency":2500}""") };
-        yield return new object[] { (new Device { DeviceType = "Handheld" }, """{"type":"device","device_type":"Handheld"}""") };
-        yield return new object[] { (new Device { BatteryStatus = "Charging" }, """{"type":"device","battery_status":"Charging"}""") };
-        yield return new object[] { (new Device { DeviceUniqueIdentifier = "d610540d-11d6-4daa-a98c-b71030acae4d" }, """{"type":"device","device_unique_identifier":"d610540d-11d6-4daa-a98c-b71030acae4d"}""") };
-        yield return new object[] { (new Device { SupportsVibration = false }, """{"type":"device","supports_vibration":false}""") };
-        yield return new object[] { (new Device { SupportsAccelerometer = true }, """{"type":"device","supports_accelerometer":true}""") };
-        yield return new object[] { (new Device { SupportsGyroscope = true }, """{"type":"device","supports_gyroscope":true}""") };
-        yield return new object[] { (new Device { SupportsAudio = true }, """{"type":"device","supports_audio":true}""") };
-        yield return new object[] { (new Device { SupportsLocationService = true }, """{"type":"device","supports_location_service":true}""") };
+        using (new AssertionScope())
+        {
+            actual.Name.Should().Be(expected.Name);
+            actual.Manufacturer.Should().Be(expected.Manufacturer);
+            actual.Brand.Should().Be(expected.Brand);
+            actual.Architecture.Should().Be(expected.Architecture);
+            actual.BatteryLevel.Should().BeApproximately(expected.BatteryLevel, Delta);
+            actual.IsCharging.Should().Be(expected.IsCharging);
+            actual.IsOnline.Should().Be(expected.IsOnline);
+            actual.BootTime.Should().Be(expected.BootTime);
+            actual.ExternalFreeStorage.Should().Be(expected.ExternalFreeStorage);
+            actual.ExternalStorageSize.Should().Be(expected.ExternalStorageSize);
+            actual.ScreenResolution.Should().Be(expected.ScreenResolution);
+            actual.ScreenDensity.Should().Be(expected.ScreenDensity);
+            actual.ScreenDpi.Should().Be(expected.ScreenDpi);
+            actual.Family.Should().Be(expected.Family);
+            actual.FreeMemory.Should().Be(expected.FreeMemory);
+            actual.FreeStorage.Should().Be(expected.FreeStorage);
+            actual.MemorySize.Should().Be(expected.MemorySize);
+            actual.Model.Should().Be(expected.Model);
+            actual.ModelId.Should().Be(expected.ModelId);
+            actual.Orientation.Should().Be(expected.Orientation);
+            actual.Simulator.Should().Be(expected.Simulator);
+            actual.StorageSize.Should().Be(expected.StorageSize);
+            actual.Timezone.Should().Be(expected.Timezone);
+            actual.UsableMemory.Should().Be(expected.UsableMemory);
+            actual.LowMemory.Should().Be(expected.LowMemory);
+            actual.ProcessorCount.Should().Be(expected.ProcessorCount);
+            actual.CpuDescription.Should().Be(expected.CpuDescription);
+            actual.ProcessorFrequency.Should().BeApproximately(expected.ProcessorFrequency, Delta);
+            actual.SupportsVibration.Should().Be(expected.SupportsVibration);
+            actual.DeviceType.Should().Be(expected.DeviceType);
+            actual.BatteryStatus.Should().Be(expected.BatteryStatus);
+            actual.DeviceUniqueIdentifier.Should().Be(expected.DeviceUniqueIdentifier);
+            actual.SupportsAccelerometer.Should().Be(expected.SupportsAccelerometer);
+            actual.SupportsGyroscope.Should().Be(expected.SupportsGyroscope);
+            actual.SupportsAudio.Should().Be(expected.SupportsAudio);
+            actual.SupportsLocationService.Should().Be(expected.SupportsLocationService);
+        }
     }
 }
